refactor: Add NumThreads to GlobalContext
This commit adds a new member variable, `mNumThreads`, to the `GlobalContext` struct in `globals.h`. The `mNumThreads` variable is used to store the number of threads to use for the compiler. Additionally, a new setter function, `GLOBAL_CONTEXT_SETTER_VALUE(NumThreads, mNumThreads)`, is added to set the value of `mNumThreads`.

A new class, `NonCopyable`, is also added to `stl.h`. This class is used to prevent copying of objects by deleting the copy constructor and assignment operator.

In `cl_args.cc`, a new command line option, `gNumThreads`, is added to specify the number of threads to use. This option is associated with the `mNumThreads` member variable of `GlobalContext`.

Finally, a new file, `threads.cc`, is added to the `common` directory. This file contains the implementation of the `GetNumThreads()` function, which returns the number of threads to use. If the `--threads` flag is set, it returns the value specified by the flag. Otherwise, it returns the number of threads available on the system.