refactor: Add Castable class for type casting
This commit adds a new class called Castable, which is used for type casting between different types. The Castable class provides methods for casting to the child type and checking if a type is the same as the given type. This class is useful when we want to encapsulate a type and cast it to its child type. The Castable class is defined in the castable.h file.

The commit also includes a new class called Node in the node.h file. The Node class represents a single unit in the abstract syntax tree (AST) and is used to organize language elements into a tree structure. The Node class inherits from the Castable class and provides methods for getting the location of the node. The Node class is part of the frontend module.

These changes improve the code structure and make it easier to work with different types and organize language elements in the AST.